일반게시판 조회 수정

+++ JSP_Project/WebContent/views/board/boardDetailView.jsp
@@ -1,8 +1,11 @@
+<%@page import="com.kh.board.model.vo.Attachment"%>
 <%@ page import="com.kh.board.model.vo.Board" %>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
 <%
 	Board b = (Board)request.getAttribute("b");
+
+	Attachment at = (Attachment)request.getAttribute("at");
 %>
 <!DOCTYPE html>
 <html>
@@ -43,9 +46,12 @@
 			<tr>
 				<th>첨부파일</th>
 				<td colspan="3">
-					첨부파일이 없을 경우 : 첨부파일이 없습니다로 표시
-					첨부파일이 있을경우 : 
-					<a download="파일원본명" href="해당파일이존재하는경로">파일원본명</a>
+					<% if(at == null){ %>
+						첨부파일이 없습니다.
+					<% } else { %>
+						<a download="<%= at.getOrignName() %>" href="<%= contextPath %>/<%= at.getFilePath()+at.getChangeName() %>"><%= at.getOrignName() %></a>
+					<% } %>
+
 				</td>
 			</tr>
 		</table>
@@ -56,10 +62,10 @@
 			<a href="<%= contextPath %>/list.bo?currentPage=1" class="btn btn-secondary btn-sm">목록가기</a>
 			
 			<!-- 로그인한 사용자가 해당 게시글의 작성자인 경우 -->
-<%-- 			<% if(loginUser != null && loginUser.getUserId().equals(b.getBoardWriter())){ %>
-				<a href="<%= contextPath %>/updateForm.bo?bno=<%= b.getBoardNo() %>" class="btn btn-warning btn-sm">수정하기</a>
+ 			<% if(loginUser != null && loginUser.getUserId().equals(b.getBoardWriter())){ %>
+				<a href="<%= contextPath %>/update.bo?bno=<%= b.getBoardNo() %>" class="btn btn-warning btn-sm">수정하기</a>
 				<a href="<%= contextPath %>/delete.bo?bno=<%= b.getBoardNo() %>" class="btn btn-danger btn-sm">삭제하기</a>
-			<%} %> --%>
+			<%} %>
 		</div>
 		
 		<br>
